refactor: trackpart mapping user 엔티티로 이동
