+++ src/features/onboarding/ui/TrackUnivStep.tsx
@@ -11,11 +11,8 @@ import { TextArea } from '@/shared/ui/text-area/TextArea';
 import { FieldGroup } from '@/shared/ui/field-group/FieldGroup';
 import { SolidButton } from '@/shared/ui/button/solid-button/SolidButton';
 import { useState } from 'react';
-import {
-  ONBOARDING_EVENTS,
-  OnBoardingFormData,
-  TrackPart,
-} from '@/features/onboarding/model/types';
+import { TrackPart } from '@/entities/user/model/types';
+import { ONBOARDING_EVENTS, OnBoardingFormData } from '@/features/onboarding/model/types';
 import { formatTrackLabel, mapToApiTrack } from '../lib/trackMapper';
 import { trackOnBoardingEvent } from '../lib/trackOnBoardingEvent';
